rename readout module
